feat: Add modules for API, menu, preview, UI, and utils
This commit adds new modules for API, menu, preview, UI, and utils. It also defines constants for the OLLAMA API URL, request timeout, name, version, authors, and description. These changes provide a foundation for implementing new features and improvements.
